refactor: Split off canvas functionality into separate file
